chore: moved facilityform logic to the correct component
